Desenvolvimento do nível 3 da prova (#3)
* Adicionado ValidationBean nos Forms do projeto

* Criado RestControllerAdvice para customizar erros de validação do BeanValidation

* Adicionado tratamento para error 404 NotFound

* Ajuste para que o RestControllerAdvice faça o tratamento dos erros de item inativo e pedido fechado

* Adicionado filtros as listagens de Pedido e Item

* Desenvolvido testes unitários para as classes Service

* Desenvolvido testes unitários para as classes Mappers

* Adicionado tratamento para quando tentar excluir um Item que já esteja em um Produto

* Desenvolvido testes unitários para as classes Controller

* Adicionado testes para casos onde o corpo requisicao é inválido

* Adicionado Swagger ao projeto

+++ src/main/java/com/seniorsistemas/project/domain/item/service/ItemServiceImpl.java
@@ -3,8 +3,11 @@ package com.seniorsistemas.project.domain.item.service;
 import java.util.Optional;
 import java.util.UUID;
 
+import com.seniorsistemas.project.config.validation.exception.ItemIsBeingUsedInPedidoException;
+import com.seniorsistemas.project.config.validation.exception.NotFoundException;
 import com.seniorsistemas.project.domain.item.dto.ItemDTO;
 import com.seniorsistemas.project.domain.item.entity.Item;
+import com.seniorsistemas.project.domain.item.entity.TipoItem;
 import com.seniorsistemas.project.domain.item.form.ItemForm;
 import com.seniorsistemas.project.domain.item.mapper.ItemMapper;
 import com.seniorsistemas.project.domain.item.repository.ItemRepository;
@@ -20,14 +23,15 @@ public class ItemServiceImpl implements ItemService {
     private ItemRepository itemRepository;
 
     @Override
-    public Page<ItemDTO> findAll(Pageable pageable) {
-        Page<Item> itemPage = itemRepository.findAll(pageable);
+    public Page<ItemDTO> findAll(String descricao, TipoItem tipo, Pageable pageable) {
+        Page<Item> itemPage = itemRepository.findAll(descricao, tipo, pageable);
         return itemPage.map(ItemMapper.MAPPER::toDTO);
     }
 
     @Override
-    public Optional<ItemDTO> findById(UUID id) {
-        return ItemMapper.MAPPER.toOptionalItemDTO(itemRepository.findById(id));
+    public ItemDTO findById(UUID id) {
+        validateNotFound(id);
+        return ItemMapper.MAPPER.toDTO(itemRepository.findById(id).get());
     }
 
     @Override
@@ -39,22 +43,36 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public ItemDTO update(ItemForm itemForm) {
+        validateNotFound(itemForm.getId());
+
         Item item = ItemMapper.MAPPER.toEntity(itemForm);
         return ItemMapper.MAPPER.toDTO(itemRepository.save(item));
     }
 
     @Override
     public void delete(UUID id) {
+        validateNotFound(id);
+        if (itemRepository.verifyIfItemIsBeingUsedInPedido(id) > 0) {
+            throw new ItemIsBeingUsedInPedidoException();
+        }
+
         itemRepository.deleteById(id);
     }
 
     @Override
     public void inactivate(UUID id) {
-        Optional<Item> optionalItem = itemRepository.findById(id);
+        validateNotFound(id);
 
-        optionalItem.ifPresent(item -> {
-            item.setAtivo(false);
-            itemRepository.save(item);
-        });
+        Item item = itemRepository.findById(id).get();
+        item.setAtivo(false);
+        itemRepository.save(item);
+    }
+
+    @Override
+    public void validateNotFound(UUID id) {
+        Optional<Item> optionalItem = itemRepository.findById(id);
+        if (optionalItem.isEmpty()) {
+            throw new NotFoundException(id);
+        }
     }
 }
